Show relevant error on posting page
When a user posts to a list he is not subscribed to, Hyperkitty tries to
subscribe him before posting. However, if the subscription policy is
set to moderate the user is not subscribed and a 500 Error is
shown. Instead show a message asking him to subscribe before posting
to the list as it is moderated.