Remove SingleAsync function and replace lockfile with flufl.lock
This commit removes out custom implementation of SingletonAsync task and removes
the decorator we added to run tasks in an asynchronous fashion. Instead, we just
use the simple AsyncTaks API to run functions with arguments without any sort of
optimization for tasks occurring more than once.

The rationale for this is instead of preventing new tasks to be added, we should
optimize the re-run of tasks. For example, a already-sorted thread should be
extremely fast to sort when the next email comes in, preventing the further need
to optimize at job scheduling level.

Also, remove the un-maintained library lockfile and use flufl.lock instead to
achieve the same functionality.