Fix the bugs in email export API.
Due to a bug, emails exported would skip the content if there was any
attachment in the Email. This fixes that bug by setting the content correctly.

Also, attachments' content are saved to a Folder when a config value,
HYPERKITTY_ATTACHMENTS_FOLDER is set. Use the get_content() API to get those
attachments instead of directly accessing the .content attribute of the API.

Fixes #252