Skip non-existent threads from the task queue.
It is possible in certain situations that a thread is deleted after a task
for it is queued in the async task queue. This results in a
Thread.DoesNotExist being raised. However, this is an expected scenario and
should not really be raising exceptions.

This fixes https://gitlab.com/mailman/hyperkitty/issues/245