Remove unintentional indent display in message view
Because .email-body preserves whitespace, the spaces in between the
end of the opening div tag and the start of the content was displayed
in the browser.

The fix, which was already present in the ajax/temp_message.html
template, which mirrors this one, is to move the > at the end of the
tag right before the content, so that no whitespace occurrs inside the
element.